+++ scripts/content.js
@@ -13,213 +13,220 @@ const getDefaultElements = () => [
   {
     id: "logo",
     selector: "//ytd-topbar-logo-renderer",
-    hidden: undefined,
+    hidden: false,
     category: "General",
   },
   {
     id: "microphone-search",
     selector: "//*[@id='voice-search-button']",
-    hidden: undefined,
+    hidden: false,
     category: "General",
   },
   {
     id: "create",
     selector:
       "(//div[@id='buttons' and contains(@class, 'ytd-masthead')]//a/../..)[1]",
-    hidden: undefined,
+    hidden: false,
     category: "General",
   },
   {
     id: "notifications",
     selector: "//ytd-notification-topbar-button-renderer",
-    hidden: undefined,
+    hidden: false,
     category: "General",
   },
   {
     id: "home",
     selector:
-      "((//ytd-guide-section-renderer)[1]//ytd-guide-entry-renderer)[1] | (//ytd-mini-guide-entry-renderer)[1]",
-    hidden: undefined,
+      "//ytd-guide-entry-renderer[a[@href='/']] | //ytd-mini-guide-entry-renderer[a[@href='/']]",
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "shorts",
     selector:
-      "((//ytd-guide-section-renderer)[1]//ytd-guide-entry-renderer)[2] | (//ytd-mini-guide-entry-renderer)[2]",
-    hidden: undefined,
+      "//ytd-guide-entry-renderer[a[@title='Shorts']] | //ytd-mini-guide-entry-renderer[a[@title='Shorts']]",
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "subscriptions",
     selector:
-      "((//ytd-guide-section-renderer)[1]//ytd-guide-entry-renderer)[3] | (//ytd-mini-guide-entry-renderer)[3]",
-    hidden: undefined,
+      "//ytd-guide-entry-renderer[a[@href='/feed/subscriptions']] | //ytd-mini-guide-entry-renderer[a[@href='/feed/subscriptions']]",
+    hidden: false,
+    category: "Sidebar",
+  },
+  {
+    id: "youtube-music",
+    selector:
+      "(//ytd-guide-entry-renderer[a[@href='https://music.youtube.com/']])[1] | //ytd-mini-guide-entry-renderer[a[@href='https://music.youtube.com/']]",
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "my-channel",
-    selector:
-      "((//ytd-guide-section-renderer)[1]//ytd-guide-entry-renderer)[5]",
-    hidden: undefined,
+    selector: "(//div[@id='section-items']/ytd-guide-entry-renderer)[1]",
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "history",
-    selector:
-      "((//ytd-guide-section-renderer)[1]//ytd-guide-entry-renderer)[6]",
-    hidden: undefined,
+    selector: "(//div[@id='section-items']/ytd-guide-entry-renderer)[2]",
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "playlists",
-    selector:
-      "((//ytd-guide-section-renderer)[1]//ytd-guide-entry-renderer)[7]",
-    hidden: undefined,
+    selector: "(//div[@id='section-items']/ytd-guide-entry-renderer)[3]",
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "my-videos",
-    selector:
-      "((//ytd-guide-section-renderer)[1]//ytd-guide-entry-renderer)[8]",
-    hidden: undefined,
+    selector: "(//div[@id='section-items']/ytd-guide-entry-renderer)[4]",
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "watch-later",
-    selector:
-      "((//ytd-guide-section-renderer)[1]//ytd-guide-entry-renderer)[9]",
-    hidden: undefined,
+    selector: "(//div[@id='section-items']/ytd-guide-entry-renderer)[5]",
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "liked-videos",
-    selector:
-      "((//ytd-guide-section-renderer)[1]//ytd-guide-entry-renderer)[10]",
-    hidden: undefined,
+    selector: "(//div[@id='section-items']/ytd-guide-entry-renderer)[6]",
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "my-clips",
+    selector: "(//div[@id='section-items']/ytd-guide-entry-renderer)[7]",
+    hidden: false,
+    category: "Sidebar",
+  },
+  {
+    id: "transfers",
     selector:
-      "((//ytd-guide-section-renderer)[1]//ytd-guide-entry-renderer)[12]",
-    hidden: undefined,
+      "//div[@id='section-items']/ytd-guide-downloads-entry-renderer  | //ytd-mini-guide-entry-renderer[a[@href='/feed/downloads']]",
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "subscriptions-panel",
     selector: "(//ytd-guide-section-renderer)[2]",
-    hidden: undefined,
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "explore-panel",
     selector: "(//ytd-guide-section-renderer)[3]",
-    hidden: undefined,
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "youtube-panel",
     selector: "(//ytd-guide-section-renderer)[4]",
-    hidden: undefined,
+    hidden: false,
     category: "Sidebar",
   },
   {
     id: "tabs",
     selector:
       "//ytd-feed-filter-chip-bar-renderer[@component-style='FEED_FILTER_CHIP_BAR_STYLE_TYPE_DEFAULT']/..",
-    hidden: undefined,
+    hidden: false,
     category: "HomePage",
   },
   {
     id: "ads",
     selector:
       "//ytd-ad-slot-renderer/ancestor::ytd-rich-item-renderer | //*[@id='player-ads'] | //ytd-banner-promo-renderer/..",
-    hidden: undefined,
+    hidden: false,
     category: "HomePage",
   },
   {
     id: "posts",
     selector: "//ytd-rich-section-renderer",
-    hidden: undefined,
+    hidden: false,
     category: "HomePage",
   },
   {
     id: "subscriptions-shorts",
     selector: "//ytd-rich-shelf-renderer/../.. | //ytd-rich-shelf-renderer",
-    hidden: undefined,
+    hidden: false,
     category: "Subscriptions",
   },
   {
     id: "subscriptions-scheduled-videos",
     selector: "//ytd-rich-item-renderer[.//ytd-toggle-button-renderer]",
-    hidden: undefined,
+    hidden: false,
     category: "Subscriptions",
   },
   {
     id: "video-title",
     selector: "//div[@id='above-the-fold']/div[@id='title']",
-    hidden: undefined,
+    hidden: false,
     category: "Video",
   },
   {
     id: "video-likes-dislikes",
     selector: "//segmented-like-dislike-button-view-model",
-    hidden: undefined,
+    hidden: false,
     category: "Video",
   },
   {
     id: "video-description",
     selector: "//div[@id='description-inner']/parent::div",
-    hidden: undefined,
+    hidden: false,
     category: "Video",
   },
   {
     id: "expand-video-description",
     selector: "//div[@id='description-inner']/parent::div",
-    hidden: undefined,
+    hidden: false,
     category: "Video",
   },
   {
     id: "video-comments",
     selector: "//ytd-comments[@id='comments']",
-    hidden: undefined,
+    hidden: false,
     category: "Video",
   },
   {
     id: "video-categories-games",
     selector: "//ytd-rich-metadata-row-renderer/../..",
-    hidden: undefined,
+    hidden: false,
     category: "Video",
   },
   {
     id: "video-ads",
     selector: "//div[@id='player-ads']",
-    hidden: undefined,
+    hidden: false,
     category: "Video",
   },
   {
     id: "video-tabs",
     selector: "//yt-related-chip-cloud-renderer",
-    hidden: undefined,
+    hidden: false,
     category: "Video",
   },
   {
     id: "video-suggested-videos",
     selector:
       "//div[@id='contents']/parent::ytd-item-section-renderer[contains(@class, 'watch-next')]",
-    hidden: undefined,
+    hidden: false,
     category: "Video",
   },
   {
     id: "video-suggested-shorts",
-    selector: undefined,
-    hidden: undefined,
+    selector: "//ytd-reel-shelf-renderer",
+    hidden: false,
     category: "Video",
   },
   {
     id: "stream-chat",
     selector: "//div[@id='chat-container']",
-    hidden: undefined,
+    hidden: false,
     category: "Stream",
   },
 ];
@@ -269,7 +276,7 @@ const setElementClick = async (target, click) => {
       null
     ).singleNodeValue;
 
-    if (elementToClick) {
+    if (click) {
       elementToClick.click();
     }
 
@@ -285,7 +292,9 @@ const setElementVisibility = async (target, hide) => {
     element.hidden = hide;
     if (target !== "expand-video-description") {
       applyVisibility(element.selector, hide);
-      waitForElements(element.selector, () => applyVisibility(element.selector, hide));
+      waitForElements(element.selector, () =>
+        applyVisibility(element.selector, hide)
+      );
     } else {
       setElementClick(target, hide);
     }
@@ -309,8 +318,12 @@ const applyVisibility = (selector, hide) => {
       elementsToHide.snapshotItem(i).style.display = hide ? "none" : "";
     }
 
-    if (selector === "stream-chat" && document.getElementById("panels-full-bleed-container")) {
-      document.getElementById("panels-full-bleed-container").style.display = hide ? "none" : "";
+    if (
+      selector === "stream-chat" &&
+      document.getElementById("panels-full-bleed-container")
+    ) {
+      document.getElementById("panels-full-bleed-container").style.display =
+        hide ? "none" : "";
     }
   });
 };
@@ -370,7 +383,10 @@ const debounce = (func, wait) => {
 const initialize = () => {
   document.addEventListener("DOMContentLoaded", applyElementVisibility);
   window.addEventListener("load", applyElementVisibility);
-  window.addEventListener("yt-navigate-finish", debounce(applyElementVisibility, 200));
+  window.addEventListener(
+    "yt-navigate-finish",
+    debounce(applyElementVisibility, 200)
+  );
   window.addEventListener("resize", debounce(applyElementVisibility, 200));
 };